requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic<2.0.0
pandas>=2.2.2
numpy>=1.26.4
tqdm>=4.66.4
transformers>=4.41.0
accelerate>=0.31.0
sentencepiece>=0.2.0
torch>=2.3.0
faiss-cpu>=1.8.0
bitsandbytes>=0.43.1; platform_system != "Windows"  # optional, skip on Windows
talib-binary>=0.4.29  # optional; or TA-Lib proper
pandas-ta>=0.3.14b
pypfopt>=1.5.5
farm-haystack[faiss]>=1.26.2  # haystack retrieval infra
vectorbt>=0.26.0
numba>=0.59.0
supabase>=2.7.4
alpaca-trade-api>=3.2.0  # optional for orders
langgraph>=0.1.0  # optional

